DeepQ_Final/requirements.txt
absl-py==0.7.1
astor==0.7.1
bleach==1.5.0
certifi==2019.3.9
chardet==3.0.4
Cython==0.29.9
gast==0.2.2
google-pasta==0.1.7
grpcio==1.16.1
h5py==2.9.0
html5lib==0.9999999
idna==2.8
kaggle==1.5.4
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keras-resnet==0.2.0
keras-retinanet==0.5.0
Markdown==3.1
mock==3.0.5
numpy==1.16.4
opencv-python==4.1.0.25
panda==0.3.1
pandas==0.24.2
Pillow==6.0.0
progressbar2==3.42.0
protobuf==3.8.0
python-dateutil==2.8.0
python-slugify==3.0.2
python-utils==2.3.0
pytz==2019.1
PyYAML==5.1
requests==2.22.0
scipy==1.2.1
six==1.12.0
tensorboard==1.9.0
tensorflow-estimator==1.13.0
tensorflow-gpu==1.9.0
termcolor==1.1.0
text-unidecode==1.2
tqdm==4.32.1
urllib3==1.24.3
Werkzeug==0.15.2
wrapt==1.11.1 
